+++ src/deploy.ts
@@ -1,182 +1,22 @@
-import dotenv from 'dotenv'
-import { Client } from 'pg'
-import path from 'path'
 import admin from 'firebase-admin'
-import { argv as args } from 'process'
-import { FunctionTypes } from './'
-
-dotenv.config()
-
-if (args.includes('--initApp')) {
-  admin.initializeApp()
-}
-
-const isDebug = args.includes('--debug')
-
-const getCliArgument = (
-  nameWithDashes: string,
-  defaultVal?: string
-): string => {
-  const arg = args.find((arg) => arg.includes(nameWithDashes))
-
-  if (arg) {
-    const val = arg.split('=')[1]
-
-    if (isDebug) {
-      console.debug(`${nameWithDashes} = "${val}"`)
-    }
-
-    return val
-  }
-
-  if (defaultVal !== undefined) {
-    if (isDebug) {
-      console.debug(`${nameWithDashes} = "${defaultVal}" (default)`)
-    }
-
-    return defaultVal
-  }
-
-  if (isDebug) {
-    console.debug(`${nameWithDashes} empty`)
-  }
-
-  return ''
-}
-
-const supabaseFunctionNamesToOnlyWorkWith = getCliArgument('--functions', '')
-  .split(',')
-  .filter((arg) => arg)
-const baseUrl = getCliArgument('--baseUrl')
-const customApiKey = getCliArgument('--customApiKey')
-
-if (!baseUrl) {
-  throw new Error('Base URL is not set!')
-}
-if (!customApiKey) {
-  throw new Error('Custom API key is not set!')
-}
-
-const pathToFirebaseFunctionsFile = path.resolve(process.cwd(), 'index.js')
-
-let postgresClient
-
-async function connectToPostgres() {
-  postgresClient = new Client(process.env.POSTGRESQL_CONNECTION_URL)
-  await postgresClient.connect()
-  return postgresClient
-}
-
-async function disconnectFromPostgres() {
-  await postgresClient.end()
-}
-
-async function runSqlQuery(query, params = []) {
-  // if (isDryRun) {
-  //   if (isDebug) console.debug(`Skipping query (dry run)`, query, params);
-  //   return;
-  // }
-
-  if (isDebug) console.debug(query, params)
-
-  return postgresClient.query(query, params)
-}
-
-function getFirebaseFunctionsAsSupabaseFunctions() {
-  const functionsByName = require(pathToFirebaseFunctionsFile)
-
-  const supabaseFunctions = {}
-
-  for (const [functionName, functionBody] of Object.entries(functionsByName)) {
-    // only accept functions created using our internal tool
-    // @ts-ignore
-    if (!functionBody._supabase) {
-      continue
-    }
-
-    // @ts-ignore
-    const { table, type } = functionBody._supabase
-
-    supabaseFunctions[functionName] = {
-      name: functionName,
-      table,
-      event: type,
-    }
-  }
-
-  return supabaseFunctions
-}
-
-const getSqlEventForEventName = (eventName) => {
-  switch (eventName) {
-    case FunctionTypes.CREATE:
-      return 'INSERT'
-    case FunctionTypes.UPDATE:
-      return 'UPDATE'
-    default:
-      throw new Error(
-        `Cannot get SQL event for event name "${eventName}": unknown!`
-      )
-  }
-}
-
-interface FunctionConfig {
-  table: string
-  event: string
-}
-
-async function createSupabaseFunctions(supabaseFunctions: {
-  [functionName: string]: FunctionConfig
-}) {
-  const output = {
-    functions: {},
-  }
-
-  for (const [functionName, functionConfig] of Object.entries(
-    supabaseFunctions
-  )) {
-    const method = 'POST'
-    const url = `${baseUrl}/${functionName}`
-    const tableName = functionConfig.table
-    const eventName = functionConfig.event
-    const headers = {
-      'content-type': 'application/json',
-      'x-api-key': customApiKey,
-    }
-
-    output.functions[functionName] = {
-      name: functionName,
-      table: tableName,
-      event: eventName,
-      type: 'http',
-      request: {
-        method,
-        url,
-      },
-      headers,
-    }
-
-    console.debug(
-      `Function "${functionName}": ${tableName}.${eventName} -> ${method} ${url}`
-    )
-
-    await runSqlQuery(
-      `DROP TRIGGER IF EXISTS ${functionName} ON public.${tableName}`
-    )
-
-    await runSqlQuery(`CREATE TRIGGER ${functionName}
-AFTER ${getSqlEventForEventName(eventName)} ON public.${tableName}
-FOR EACH ROW
-EXECUTE PROCEDURE supabase_functions.http_request('${url}', '${method}', '${JSON.stringify(
-      headers
-    )}', '{}', '1000');`)
-  }
-}
+import { getCliArgs } from './env'
+import {
+  connectToPostgres,
+  createSupabaseFunctions,
+  disconnectFromPostgres,
+  getFirebaseFunctionsAsSupabaseFunctions,
+} from './sql'
 
 async function main() {
   try {
-    console.info('Starting up!')
-    console.info(`Using firebase functions file ${pathToFirebaseFunctionsFile}`)
+    console.info('Starting up...')
+
+    const { args, functionNames } = getCliArgs()
+
+    if (args.initApp) {
+      console.debug(`initializing firebase admin app...`)
+      admin.initializeApp()
+    }
 
     const allSupabaseFunctions = getFirebaseFunctionsAsSupabaseFunctions()
 
@@ -186,22 +26,18 @@ async function main() {
       )
     }
 
-    console.debug(
+    console.info(
       `Found ${Object.keys(allSupabaseFunctions).length} firebase functions`
     )
 
     let supabaseFunctionsToOperateOn = allSupabaseFunctions
 
-    if (supabaseFunctionNamesToOnlyWorkWith.length) {
-      console.debug(
-        `Only operating on functions: ${supabaseFunctionNamesToOnlyWorkWith.join(
-          ', '
-        )}`
-      )
+    if (functionNames !== null && functionNames.length) {
+      console.debug(`Only operating on functions: ${functionNames.join(', ')}`)
 
       const newFuncs = {}
 
-      for (const functionName of supabaseFunctionNamesToOnlyWorkWith) {
+      for (const functionName of functionNames) {
         if (!(functionName in allSupabaseFunctions)) {
           throw new Error(
             `Cannot operate on function "${functionName}": does not exist!`
@@ -213,13 +49,15 @@ async function main() {
       supabaseFunctionsToOperateOn = newFuncs
     }
 
-    console.debug('Connecting to PostgreSQL...')
+    console.info('Connecting to PostgreSQL...')
 
     await connectToPostgres()
 
-    console.debug('Connected!')
+    console.info('Connected!')
 
-    console.info(`Functions will be called via URL ${baseUrl}/myFunctionName`)
+    console.info(
+      `Functions will be called via URL ${args.baseUrl}/myFunctionName`
+    )
 
     await createSupabaseFunctions(supabaseFunctionsToOperateOn)